refactor(plugin-flow-builder): renaming and create content-fields as a module folder
