+++ packages/botonic-plugin-flow-builder/src/content-fields/flow-element.tsx
@@ -1,6 +1,6 @@
 import { CarouselElementNode } from '../flow-builder-models'
-import { FlowButton } from './button'
-import { ContentFieldsBase } from './content-base'
+import { ContentFieldsBase } from './content-fields-base'
+import { FlowButton } from './flow-button'
 
 export class FlowElement extends ContentFieldsBase {
   public title = ''
