+++ packages/botonic-plugin-flow-builder/src/content-fields/types.ts
@@ -1,0 +1,5 @@
+import { FlowCarousel } from './flow-carousel'
+import { FlowImage } from './flow-image'
+import { FlowText } from './flow-text'
+
+export type FlowContent = FlowText | FlowImage | FlowCarousel
